Refactor date formatting logic to use PascalCase for enum values and update import paths
