+++ src/shared/enums/date-type.enum.ts
@@ -1,7 +1,7 @@
 export const enum DateFormatType {
-  long = 'long',
-  long_with_year = 'long_with_year',
-  short = 'short',
-  short_with_year = 'short_with_year',
-  time_only = 'time_only',
+  Long = 'Long',
+  LongWithYear = 'LongWithYear',
+  Short = 'Short',
+  ShortWithYear = 'ShortWithYear',
+  TimeOnly = 'TimeOnly',
 }
