+++ src/shared/utils/format-date.util.ts
@@ -1,4 +1,4 @@
-import { DateFormatType } from '../../shared/enums/date-type.enum';
+import { DateFormatType } from '../enums/date-type.enum';
 
 export function FormatDate(
   input: Date | string | number,
@@ -9,7 +9,7 @@ export function FormatDate(
   if (isNaN(date.getTime())) return '';
 
   switch (type) {
-    case DateFormatType.long:
+    case DateFormatType.Long:
       return new Intl.DateTimeFormat(locale, {
         day: 'numeric',
         month: 'long',
@@ -17,7 +17,7 @@ export function FormatDate(
         minute: '2-digit',
       }).format(date);
 
-    case DateFormatType.long_with_year:
+    case DateFormatType.LongWithYear:
       return new Intl.DateTimeFormat(locale, {
         day: 'numeric',
         month: 'long',
@@ -26,20 +26,20 @@ export function FormatDate(
         minute: '2-digit',
       }).format(date);
 
-    case DateFormatType.short:
+    case DateFormatType.Short:
       return new Intl.DateTimeFormat(locale, {
         day: 'numeric',
         month: 'short',
       }).format(date);
 
-    case DateFormatType.short_with_year:
+    case DateFormatType.ShortWithYear:
       return new Intl.DateTimeFormat(locale, {
         day: 'numeric',
         month: 'short',
         year: 'numeric',
       }).format(date);
 
-    case DateFormatType.time_only:
+    case DateFormatType.TimeOnly:
       return new Intl.DateTimeFormat(locale, {
         hour: '2-digit',
         minute: '2-digit',
