+++ src/shared/сomponents/transform-date/transform-date.tsx
@@ -1,7 +1,7 @@
-import { FormatDate } from '../../../data-access/date-service/date.service';
 import React, { useMemo } from 'react';
 import { TransformDateProps } from '../models/transform-date.model';
 import { defaultLocale } from '../../utils/get-preferred-locale.util';
+import { FormatDate } from '../../utils/format-date.util';
 
 export default function TransformDate({ date, type, locale = defaultLocale }: TransformDateProps) {
   const formatted = useMemo(() => {
